change loop iteration variable to size_t
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@3385 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/datadriven/tools/DatasetGenerator.cpp
@@ -74,7 +74,7 @@ namespace sg {
         classes.set(i, cls);
       }
 
-      for (int i = 0; i < 4; i++) {
+      for (size_t i = 0; i < 4; i++) {
         trData.normalizeDimension(i);
       }
     }
@@ -99,7 +99,7 @@ namespace sg {
         classes.set(i, cls);
       }
 
-      for (int i = 0; i < 4; i++) {
+      for (size_t i = 0; i < 4; i++) {
         trData.normalizeDimension(i);
       }
     }
